+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.7.1 (2021-11-3)
 * Optimized Performance
+* Added a Status Bar Widget to Switch Terminal
 
 ### 0.7.0 (2021-10-26)
 * Added Support for WSL on Windows